requirements.txt
pwntools
pycryptodomex